refactor: 예외 패키지 수정
