+++ src/main/java/org/ahpuh/surf/common/exception/like/DuplicatedLikeException.java
@@ -1,4 +1,4 @@
-package org.ahpuh.surf.common.exception.post;
+package org.ahpuh.surf.common.exception.like;
 
 import org.ahpuh.surf.common.exception.ApplicationException;
 import org.ahpuh.surf.common.exception.ExceptionType;
