+++ src/main/java/org/ahpuh/surf/user/domain/User.java
@@ -4,9 +4,9 @@ import lombok.*;
 import org.ahpuh.surf.category.domain.Category;
 import org.ahpuh.surf.common.domain.BaseEntity;
 import org.ahpuh.surf.common.exception.category.DuplicatedCategoryException;
-import org.ahpuh.surf.common.exception.post.DuplicatedLikeException;
+import org.ahpuh.surf.common.exception.follow.DuplicatedFollowingException;
+import org.ahpuh.surf.common.exception.like.DuplicatedLikeException;
 import org.ahpuh.surf.common.exception.post.DuplicatedPostException;
-import org.ahpuh.surf.common.exception.user.DuplicatedFollowingException;
 import org.ahpuh.surf.common.exception.user.InvalidPasswordException;
 import org.ahpuh.surf.follow.domain.Follow;
 import org.ahpuh.surf.like.domain.Like;
